+++ database/1-social-counts.sql
@@ -1,0 +1,5 @@
+ALTER TABLE public.entries ADD stumbleupon INT NULL;
+ALTER TABLE public.entries ADD pinterest INT NULL;
+ALTER TABLE public.entries ADD linkedin INT NULL;
+ALTER TABLE public.entries ADD facebook INT NULL;
+ALTER TABLE public.entries ADD sharedcount_stats JSONB NULL;