+++ docker/build.local.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+
+npm run mailer:local:build
+npm run collector:local:build
+npm run social:local:build